Fix bug

+++ vm64.h
@@ -23,7 +23,7 @@ void *get_vmx_region();
 void *get_vmcs_region();
 void *get_current_vmcs();
 void init_vm64_data_structures();
-unsigned int read_IA32_VMX_BASIC();
+unsigned long read_IA32_VMX_BASIC();
 unsigned int read_vmcs_revision_identifier();
 int vmx_clear_vmcs(void *vmcs_region);
 int vmx_enable_vmcs_region(void *vmcs_region);
